refactor(#85): refactor folder architecture
refactor for various packages from vanilla
