+++ packages/vanilla/.npmignore
