+++ packages/vanilla/package.json
@@ -1,0 +1,53 @@
+{
+    "name": "im-sang-tae",
+    "version": "1.0.3",
+    "description": "state management library that are not dependent on the framework",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "private": false,
+    "type": "module",
+    "scripts": {
+      "types": "tsc --emitDeclarationOnly",
+      "build": "npm run types && rollup -c",
+      "start": "rollup -c -w",
+      "test": "jest",
+      "start:localTest": "cd src/test && npm run start",
+      "build:localTest": "cd src/test && npm run build",
+      "build:examples": "for d in ./examples/*/ ; do (cd \"$d\" && npm run build); done",
+      "install:examples": "for d in ./examples/*/ ; do (cd \"$d\" && npm install); done",
+      "update:examples": "for d in ./examples/*/ ; do (cd \"$d\" && npm update); done"
+    },
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/modern-agile-team/im-sang-tae.git"
+    },
+    "keywords": [
+      "state",
+      "global",
+      "manager",
+      "javascript",
+      "frontend"
+    ],
+    "author": "",
+    "license": "MIT",
+    "bugs": {
+      "url": "https://github.com/modern-agile-team/im-sang-tae/issues",
+      "email": "alstnsrl98@gmail.com"
+    },
+    "homepage": "https://github.com/modern-agile-team/im-sang-tae#readme",
+    "devDependencies": {
+      "@rollup/plugin-typescript": "^11.1.1",
+      "@types/jest": "^29.5.1",
+      "jest": "^29.5.0",
+      "rollup": "^3.23.0",
+      "rollup-plugin-delete": "^2.0.0",
+      "@rollup/plugin-terser": "^0.4.3",
+      "rollup-plugin-dts": "^5.3.0",
+      "ts-jest": "^29.1.0",
+      "tslib": "^2.5.2",
+      "@types/node": "^20.1.1",
+      "ts-node": "^10.9.1",
+      "typescript": "^5.0.4"
+    }
+  }
+
