+++ packages/vanilla/rollup.config.js
