+++ packages/vanilla/src/index.ts
