+++ packages/vanilla/src/stateManager/index.ts
