+++ packages/vanilla/src/stateManager/stateManager.ts
