+++ packages/vanilla/src/store/index.ts
