+++ packages/vanilla/src/test/package.json
