+++ packages/vanilla/src/test/tsconfig.json
