+++ packages/vanilla/src/test/webpack.config.js
