+++ packages/vanilla/src/types/stateManager.d.ts
