+++ packages/vanilla/src/types/store.d.ts
